WIP: Ignore FFV positions in gunner and observer positions
Some mods appear to define FFV positions that are also other positions
(eg a gunner and an observer that can shoot when turned out). This means
fullCrew and similar functions return more positions than exist for a
vehicle.

This might mess with 'properly' defined vehicles, testing is needed.

Is there a way to check if a position is occupying the same position as
another category of crew position?